requirements/common.txt
Django>=1.5
django-suit>=0.2.2
django-suit-redactor>=0.0.2
six>=1.3
gunicorn>=0.17.2
